@@ -69,7 +69,7 @@ export 'schemas/enums/vertical_direction_schema.dart';
 export 'schemas/enums/visual_density_schema.dart';
 export 'schemas/enums/wrap_alignment_schema.dart';
 export 'schemas/enums/wrap_cross_alignment_schema.dart';
-
+//
 export 'schemas/objects/alignment_schema.dart';
 export 'schemas/objects/app_bar_theme_schema.dart';
 export 'schemas/objects/border_radius_schema.dart';
@@ -99,11 +99,15 @@ export 'schemas/objects/drawer_theme_data_schema.dart';
 export 'schemas/objects/edge_insets_geometry_schema.dart';
 export 'schemas/objects/elevated_button_theme_data_schema.dart';
 export 'schemas/objects/expansion_tile_theme_data_schema.dart';
+export 'schemas/objects/filled_button_theme_data_schema.dart';
 export 'schemas/objects/floating_action_button_theme_data_schema.dart';
 export 'schemas/objects/font_feature_schema.dart';
+export 'schemas/objects/font_variation_schema.dart';
 export 'schemas/objects/gradient_schema.dart';
 export 'schemas/objects/gradient_transform_schema.dart';
+export 'schemas/objects/icon_button_theme_data_schema.dart';
 export 'schemas/objects/icon_data_schema.dart';
+export 'schemas/objects/icon_schema.dart';
 export 'schemas/objects/icon_theme_data_schema.dart';
 export 'schemas/objects/image_provider_schema.dart';
 export 'schemas/objects/input_border_schema.dart';
@@ -123,6 +127,10 @@ export 'schemas/objects/material_state_property_outlined_border_schema.dart';
 export 'schemas/objects/material_state_property_size_schema.dart';
 export 'schemas/objects/material_state_property_text_style_schema.dart';
 export 'schemas/objects/matrix4_schema.dart';
+export 'schemas/objects/menu_bar_theme_data_schema.dart';
+export 'schemas/objects/menu_button_theme_data_schema.dart';
+export 'schemas/objects/menu_style_schema.dart';
+export 'schemas/objects/menu_theme_data_schema.dart';
 export 'schemas/objects/mouse_cursor_schema.dart';
 export 'schemas/objects/navigation_bar_theme_data_schema.dart';
 export 'schemas/objects/navigation_rail_theme_data_schema.dart';
@@ -162,6 +170,5 @@ export 'schemas/objects/text_theme_schema.dart';
 export 'schemas/objects/theme_data_schema.dart';
 export 'schemas/objects/time_picker_theme_data_schema.dart';
 export 'schemas/objects/toggle_buttons_theme_data_schema.dart';
-export 'schemas/objects/toolbar_options_schema.dart';
 export 'schemas/objects/tooltip_theme_data_schema.dart';
 export 'schemas/objects/typography_schema.dart';
