@@ -52,6 +52,7 @@ class Schemas {
     EdgeInsetsGeometrySchema.id: EdgeInsetsGeometrySchema.schema,
     ElevatedButtonThemeDataSchema.id: ElevatedButtonThemeDataSchema.schema,
     ExpansionTileThemeDataSchema.id: ExpansionTileThemeDataSchema.schema,
+    FilledButtonThemeDataSchema.id: FilledButtonThemeDataSchema.schema,
     FilterQualitySchema.id: FilterQualitySchema.schema,
     FlexFitSchema.id: FlexFitSchema.schema,
     FloatingActionButtonAnimatorSchema.id:
@@ -65,9 +66,12 @@ class Schemas {
     FontFeatureSchema.id: FontFeatureSchema.schema,
     FontStyleSchema.id: FontStyleSchema.schema,
     FontWeightSchema.id: FontWeightSchema.schema,
+    FontVariationSchema.id: FontVariationSchema.schema,
     GradientSchema.id: GradientSchema.schema,
     GradientTransformSchema.id: GradientTransformSchema.schema,
     HitTestBehaviorSchema.id: HitTestBehaviorSchema.schema,
+    IconButtonThemeDataSchema.id: IconButtonThemeDataSchema.schema,
+    IconSchema.id: IconSchema.schema,
     IconDataSchema.id: IconDataSchema.schema,
     IconThemeDataSchema.id: IconThemeDataSchema.schema,
     ImageProviderSchema.id: ImageProviderSchema.schema,
@@ -105,6 +109,10 @@ class Schemas {
     MaterialTypeSchema.id: MaterialTypeSchema.schema,
     Matrix4Schema.id: Matrix4Schema.schema,
     MaxLengthEnforcementSchema.id: MaxLengthEnforcementSchema.schema,
+    MenuBarThemeDataSchema.id: MenuBarThemeDataSchema.schema,
+    MenuButtonThemeDataSchema.id: MenuButtonThemeDataSchema.schema,
+    MenuStyleSchema.id: MenuStyleSchema.schema,
+    MenuThemeDataSchema.id: MenuThemeDataSchema.schema,
     MouseCursorSchema.id: MouseCursorSchema.schemas,
     NavigationBarThemeDataSchema.id: NavigationBarThemeDataSchema.schema,
     NavigationDestinationLabelBehaviorSchema.id:
@@ -179,7 +187,6 @@ class Schemas {
     TileModeSchema.id: TileModeSchema.schema,
     TimePickerThemeDataSchema.id: TimePickerThemeDataSchema.schema,
     ToggleButtonsThemeDataSchema.id: ToggleButtonsThemeDataSchema.schema,
-    ToolbarOptionsSchema.id: ToolbarOptionsSchema.schema,
     TooltipThemeDataSchema.id: TooltipThemeDataSchema.schema,
     TooltipTriggerModeSchema.id: TooltipTriggerModeSchema.schema,
     TypographySchema.id: TypographySchema.schema,
