@@ -11,18 +11,24 @@ class NavigationBarThemeDataSchema {
         'https://api.flutter.dev/flutter/material/NavigationBarThemeData-class.html',
     'title': 'NavigationBarThemeData',
     'type': 'object',
+    'additionalProperties': false,
     'properties': {
       'backgroundColor': SchemaHelper.objectSchema(ColorSchema.id),
       'elevation': SchemaHelper.numberSchema,
       'height': SchemaHelper.numberSchema,
       'iconTheme': SchemaHelper.objectSchema(
-          MaterialStatePropertyIconThemeDataSchema.id),
+        MaterialStatePropertyIconThemeDataSchema.id,
+      ),
       'indicatorColor': SchemaHelper.objectSchema(ColorSchema.id),
       'indicatorShape': SchemaHelper.objectSchema(ShapeBorderSchema.id),
       'labelBehavior': SchemaHelper.objectSchema(
-          NavigationDestinationLabelBehaviorSchema.id),
-      'labelTextStyle':
-          SchemaHelper.objectSchema(MaterialStatePropertyTextStyleSchema.id)
+        NavigationDestinationLabelBehaviorSchema.id,
+      ),
+      'labelTextStyle': SchemaHelper.objectSchema(
+        MaterialStatePropertyTextStyleSchema.id,
+      ),
+      'shadowColor': SchemaHelper.objectSchema(ColorSchema.id),
+      'surfaceTintColor': SchemaHelper.objectSchema(ColorSchema.id),
     },
   };
 }
