@@ -17,9 +17,13 @@ class SliderThemeDataSchema {
       'activeTrackColor': SchemaHelper.objectSchema(ColorSchema.id),
       'disabledActiveTickMarkColor': SchemaHelper.objectSchema(ColorSchema.id),
       'disabledActiveTrackColor': SchemaHelper.objectSchema(ColorSchema.id),
-      'disabledInactiveTickMarkColor':
-          SchemaHelper.objectSchema(ColorSchema.id),
+      'disabledInactiveTickMarkColor': SchemaHelper.objectSchema(
+        ColorSchema.id,
+      ),
       'disabledInactiveTrackColor': SchemaHelper.objectSchema(ColorSchema.id),
+      'disabledSecondaryActiveTrackColor': SchemaHelper.objectSchema(
+        ColorSchema.id,
+      ),
       'disabledThumbColor': SchemaHelper.objectSchema(ColorSchema.id),
       'inactiveTickMarkColor': SchemaHelper.objectSchema(ColorSchema.id),
       'inactiveTrackColor': SchemaHelper.objectSchema(ColorSchema.id),
@@ -30,23 +34,32 @@ class SliderThemeDataSchema {
       'overlappingShapeStrokeColor': SchemaHelper.objectSchema(ColorSchema.id),
       'overlayColor': SchemaHelper.objectSchema(ColorSchema.id),
       'overlayShape': SchemaHelper.objectSchema(SliderComponentShapeSchema.id),
-      'rangeThumbShape':
-          SchemaHelper.objectSchema(RangeSliderThumbShapeSchema.id),
-      'rangeTickMarkShape':
-          SchemaHelper.objectSchema(RangeSliderTickMarkShapeSchema.id),
-      'rangeTrackShape':
-          SchemaHelper.objectSchema(RangeSliderTrackShapeSchema.id),
-      'rangeValueIndicatorShape':
-          SchemaHelper.objectSchema(RangeSliderValueIndicatorShapeSchema.id),
-      'showValueIndicator':
-          SchemaHelper.objectSchema(ShowValueIndicatorSchema.id),
+      'rangeThumbShape': SchemaHelper.objectSchema(
+        RangeSliderThumbShapeSchema.id,
+      ),
+      'rangeTickMarkShape': SchemaHelper.objectSchema(
+        RangeSliderTickMarkShapeSchema.id,
+      ),
+      'rangeTrackShape': SchemaHelper.objectSchema(
+        RangeSliderTrackShapeSchema.id,
+      ),
+      'rangeValueIndicatorShape': SchemaHelper.objectSchema(
+        RangeSliderValueIndicatorShapeSchema.id,
+      ),
+      'secondaryActiveTrackColor': SchemaHelper.objectSchema(
+        ColorSchema.id,
+      ),
+      'showValueIndicator': SchemaHelper.objectSchema(
+        ShowValueIndicatorSchema.id,
+      ),
       'thumbColor': SchemaHelper.objectSchema(ColorSchema.id),
       'thumbShape': SchemaHelper.objectSchema(SliderComponentShapeSchema.id),
       'trackHeight': SchemaHelper.numberSchema,
       'trackShape': SchemaHelper.objectSchema(SliderTrackShapeSchema.id),
       'valueIndicatorColor': SchemaHelper.objectSchema(ColorSchema.id),
-      'valueIndicatorShape':
-          SchemaHelper.objectSchema(SliderComponentShapeSchema.id),
+      'valueIndicatorShape': SchemaHelper.objectSchema(
+        SliderComponentShapeSchema.id,
+      ),
       'valueIndicatorTextStyle': SchemaHelper.objectSchema(TextStyleSchema.id),
     },
   };
